fix type for math dialect (#626)

+++ src/kirin/dialects/math/stmts.py
@@ -72,7 +72,7 @@ class ceil(ir.Statement):
     name = "ceil"
     traits = frozenset({ir.Pure(), lowering.FromPythonCall()})
     x: ir.SSAValue = info.argument(types.Float)
-    result: ir.ResultValue = info.result(types.Float)
+    result: ir.ResultValue = info.result(types.Int)
 
 
 @statement(dialect=dialect)
@@ -173,7 +173,7 @@ class floor(ir.Statement):
     name = "floor"
     traits = frozenset({ir.Pure(), lowering.FromPythonCall()})
     x: ir.SSAValue = info.argument(types.Float)
-    result: ir.ResultValue = info.result(types.Float)
+    result: ir.ResultValue = info.result(types.Int)
 
 
 @statement(dialect=dialect)
@@ -356,7 +356,7 @@ class trunc(ir.Statement):
     name = "trunc"
     traits = frozenset({ir.Pure(), lowering.FromPythonCall()})
     x: ir.SSAValue = info.argument(types.Float)
-    result: ir.ResultValue = info.result(types.Float)
+    result: ir.ResultValue = info.result(types.Int)
 
 
 @statement(dialect=dialect)
